refactor: Split current-account grpc_service.go into logical modules
The 2151-line grpc_service.go was difficult to navigate and maintain.
Split into 7 focused files, all under 600 lines:

- grpc_service.go (370): Core struct, interfaces, constants, constructors
- grpc_account_endpoints.go (259): Create, Retrieve, Update account
- grpc_deposit_endpoints.go (229): ExecuteDeposit with idempotency
- grpc_withdrawal_execute.go (432): ExecuteWithdrawal, outbox completion
- grpc_withdrawal_manage.go (397): Initiate, Update, Retrieve withdrawal
- grpc_control_endpoints.go (400): Freeze/Unfreeze/Close, events, webhooks
- grpc_mappers.go (159): Proto/domain conversion helpers

No public API changes. All files remain in the same package.
